Fixed dispatcher
Initialize `events` to 0 when creating a new `poll_handler`

dispatch() was passing `events` to the handler instead of `revents`
`events` is for telling `poll()` which events to look for, `revents`
is the actual result of the poll.

`set_events()` sets `pollfd->events` to `handler->events` but
`dispatcher_flags` was setting `fds[].events` so `poll_handler.events`
would contain garbage when passed to poll. Made `dispatcher_flags` set
`handler->events` instead.